+++ test/regression/BalanceDeltaSettlement.t.sol
@@ -99,6 +99,7 @@ contract MockJBController_BalanceDelta {
             ownerMustSendPayouts: false,
             holdFees: false,
             scopeCashOutsToLocalBalances: true,
+            pauseCrossProjectFeeFreeInflows: false,
             useDataHookForPay: false,
             useDataHookForCashOut: false,
             dataHook: address(0),